+++ App.tsx
@@ -11,7 +11,6 @@ import { NavigationContainer, StatusBar, Ionicons } from './components/elements'
 import { setupDatabase } from './database/database';
 import Home from './views/home';
 import Settings from './views/settings';
-import { NavigatorScreenParams, ParamListBase } from '@react-navigation/native';
 import { createMergeableStore } from 'tinybase/mergeable-store';
 
 // import { sync } from './database/synchronize';
@@ -30,18 +29,6 @@ import { createMergeableStore } from 'tinybase/mergeable-store';
 //     supabase.auth.stopAutoRefresh()
 //   }
 // })
-
-export interface TabParamList extends ParamListBase {
-  Home: NavigatorScreenParams<HomeStackParamList>;
-  Settings: undefined;
-};
-
-export interface HomeStackParamList extends ParamListBase {
-  Index: undefined;
-  EditCar: undefined;
-  Records: undefined;
-  EditRecord: undefined;
-};
 
 init({
   dsn: 'https://ec4adae5dfe85a00b368745227de8d66@o4509037304807424.ingest.us.sentry.io/4509037306707968',
